Move 2PC to crate::two_phase_commit
This moves the two_phase_commit module to the top-level of the crate;
shortening use statements and dedicating the algorithm module to
interface implementation.
