+++ libaugrim/src/two_phase_commit/participant_algorithm.rs
